Implement judge linefeed (#29)
* Add extension class of String.

This commit is to add StringExtension.cs file which implement the method
of common String extension. It is not implemented yet.

* Rename class to JudgementExtension

This commit is to rename the class StringExtension to JudgementExtension,
because it is implemented feature for any judgement method.

* Implement linefeed judgement for Windows style.

* Implement linefeed judgement for Unix style

* Implement linefeed judgement for mac style

* Fix bug of mixing line feed

This commit is to fix the bug which return wrong value
when mixing line feed between unix style and mac ox style
at calling IsUnixStyleLineFeed method.